check if the plugin option is enabled

+++ Modules/Checkout.php
@@ -19,43 +19,47 @@ class Checkout {
 	}
 
 	public function checkout() {
-		woocommerce_form_field( 'wms-opt-in', array(
-			'type'        => 'checkbox',
-			'class'       => array( 'form-row marketing-opt-in' ),
-			'label_class' => array( 'woocommerce-form__label woocommerce-form__label-for-checkbox checkbox' ),
-			'input_class' => array( 'woocommerce-form__input woocommerce-form__input-checkbox input-checkbox' ),
-			'required'    => false,
-			'label'       => 'Marketing Opt-In',
-		) );
+		if ( get_option( 'wms_enable' ) === 'yes' ) {
+			woocommerce_form_field( 'wms-opt-in', array(
+				'type'        => 'checkbox',
+				'class'       => array( 'form-row marketing-opt-in' ),
+				'label_class' => array( 'woocommerce-form__label woocommerce-form__label-for-checkbox checkbox' ),
+				'input_class' => array( 'woocommerce-form__input woocommerce-form__input-checkbox input-checkbox' ),
+				'required'    => false,
+				'label'       => get_option( 'wms_label', __( 'I would like to sign up to receive email updates from Deedy.', 'wms' ) ),
+			), true );
+		}
 	}
 
 	public function process() {
-		$user_id = get_current_user_id();
-		if ( $user_id ) {
-			$user       = get_user_by( 'ID', $user_id );
-			$email      = sanitize_email( $user->user_email );
-			$first_name = sanitize_text_field( $user->first_name );
-			$last_name  = sanitize_text_field( $user->last_name );
-		} else {
-			if ( isset( $_POST['billing_email'] ) && is_email( $_POST['billing_email'] ) ) {
-				$email      = sanitize_email( $_POST['billing_email'] );
-				$first_name = isset( $_POST['billing_first_name'] ) ? sanitize_text_field( $_POST['billing_first_name'] ) : '';
-				$last_name  = isset( $_POST['billing_first_name'] ) ? sanitize_text_field( $_POST['billing_last_name'] ) : '';
-				$user_id    = null;
+		if ( get_option( 'wms_enable' ) === 'yes' ) {
+			$user_id = get_current_user_id();
+			if ( $user_id ) {
+				$user       = get_user_by( 'ID', $user_id );
+				$email      = sanitize_email( $user->user_email );
+				$first_name = sanitize_text_field( $user->first_name );
+				$last_name  = sanitize_text_field( $user->last_name );
 			} else {
-				return;
+				if ( isset( $_POST['billing_email'] ) && is_email( $_POST['billing_email'] ) ) {
+					$email      = sanitize_email( $_POST['billing_email'] );
+					$first_name = isset( $_POST['billing_first_name'] ) ? sanitize_text_field( $_POST['billing_first_name'] ) : '';
+					$last_name  = isset( $_POST['billing_first_name'] ) ? sanitize_text_field( $_POST['billing_last_name'] ) : '';
+					$user_id    = null;
+				} else {
+					return;
+				}
 			}
-		}
 
-		$submission = new Submissions( $email, $user_id, $first_name, $last_name );
+			$submission = new Submissions( $email, $user_id, $first_name, $last_name );
 
-		if ( isset( $_POST['wms-opt-in'] ) && $_POST['wms-opt-in'] === '1' ) {
-			if ( ! $submission->is_submission_exist() ) {
-				$submission->add_submission();
-			}
-		} else {
-			if ( $submission->is_submission_exist() ) {
-				$submission->delete_submission();
+			if ( isset( $_POST['wms-opt-in'] ) && $_POST['wms-opt-in'] === '1' ) {
+				if ( ! $submission->is_submission_exist() ) {
+					$submission->add_submission();
+				}
+			} else {
+				if ( $submission->is_submission_exist() ) {
+					$submission->delete_submission();
+				}
 			}
 		}
 	}
